pyproject.toml
[project]
name = "clash-royale-tracker"
version = "0.1.0"
description = "Track Clash Royale player stats with historical trends"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.0",
    "requests>=2.32.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
